feat: move aux function to utils
